requirements.txt
pyowm
textual
requests
schedule
python-decouple
